新增定时任务管理器   时间误差在20毫秒以内

+++ src/main/java/com/iteaj/util/module/oauth2/AuthorizeStorageManager.java
@@ -13,11 +13,11 @@ public interface AuthorizeStorageManager {
     String STORAGE_MANAGER = AuthorizeStorageManager.class.getName();
 
     /**
-     * 返回一个超时时间 毫秒单位
+     * 返回一个超时时间 秒单位
      * 超过这个时间上下文就会被清除掉
      * @return
      */
-    long getTimeout();
+    int getTimeout();
 
     /**
      * 返回一个存储上下文
